# src/main/resources/openapi/aequitas-api.yaml
openapi: 3.0.3
info:
  title: Aequitas Central Service API
  description: |
    REST API for managing time entries and user profiles within a multi-tenant legal timekeeping SaaS platform.
    
    This API follows hexagonal architecture principles and enforces tenant-level data isolation.
    All operations are scoped to the authenticated tenant context.
    
    ## Authentication
    All endpoints require authentication via OAuth2/JWT tokens. The security context provides
    tenant identification and role-based access control.
    
    ## Idempotency
    Mutation endpoints support optional `Idempotency-Key` headers to prevent duplicate operations
    caused by client retries. Keys should be UUIDs and are scoped to tenant + operation type.
    
    ## Pagination
    List endpoints use cursor-based pagination for stable iteration over large result sets.
  version: 0.1.0
  contact:
    name: Aequitas Engineering
    url: https://example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.aequitas.example.com
    description: Production server
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - User Profiles
      summary: Get authenticated user's profile
      description: |
        Retrieves the authenticated user's own profile, including identity, role,
        tenant membership, and contact information.
        
        Accessible by all authenticated users regardless of role.
      operationId: getCurrentUserProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized - missing or invalid authentication
      security:
        - oauth2: []
  /api/v1/users:
    get:
      tags:
        - User Profiles
      summary: List user profiles
      description: |
        Lists all user profiles within the authenticated tenant's firm,
        optionally filtered by role.
        
        Only accessible by users with MANAGER or ADMIN roles.
      operationId: listUserProfiles
      parameters:
        - name: role
          in: query
          description: Filter by user role
          required: false
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: User profiles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfileResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid authentication
        '403':
          description: Forbidden - insufficient permissions (requires manager/admin role)
      security:
        - oauth2: []
  /api/v1/users/{id}/role:
    patch:
      tags:
        - User Profiles
      summary: Update user role
      description: |
        Updates a user's role assignment within the tenant.
        
        Only accessible by users with ADMIN role. Cannot demote the last ADMIN.
      operationId: updateUserRole
      parameters:
        - name: id
          in: path
          description: User identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '204':
          description: User role updated successfully
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid authentication
        '403':
          description: Forbidden - insufficient permissions (requires admin role)
        '404':
          description: User not found
        '409':
          description: Cannot demote last admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - oauth2: []
...
components:
  schemas:
...
    UserProfileResponse:
      type: object
      required:
        - id
        - firmId
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
          description: User identifier
        firmId:
          type: string
          format: uuid
          description: Tenant identifier
        email:
          type: string
          format: email
          description: Email address
        role:
          $ref: '#/components/schemas/Role'
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        firmId: "660e8400-e29b-41d4-a716-446655440001"
        email: "john.doe@example.com"
        role: "EMPLOYEE"
    UpdateUserRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          $ref: '#/components/schemas/Role'
      example:
        role: "MANAGER"
...
    Role:
      type: string
      enum:
        - EMPLOYEE
        - MANAGER
        - ADMIN
      description: |
        RBAC role within a firm:
        * `EMPLOYEE` - Individual contributor with access to own entries only
        * `MANAGER` - People lead who may manage subordinate time entries
        * `ADMIN` - Firm administrator with full access inside the tenant
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status message
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: Request path that caused the error
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors (for 400 responses)
      example:
        timestamp: "2025-11-15T10:30:00Z"
        status: 400
        error: "Bad Request"
        message: "Validation failed"
        path: "/api/v1/entries"
        errors:
          - field: "narrative"
            message: "must not be blank"
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth2 authentication with JWT tokens
      flows:
        authorizationCode:
          authorizationUrl: https://auth.aequitas.example.com/oauth/authorize
          tokenUrl: https://auth.aequitas.example.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
security:
  - oauth2: []